fix(engines): restore googlevision.py and add google_vision compatibility shim
- Restore googlevision.py as the canonical module for backwards compatibility
  with downstream wrappers and configs.
- Provide google_vision.py as a compatibility shim re-exporting GoogleVisionHelper
  so both 'googlevision' and 'google_vision' config keys and import paths work.
- Add unit tests covering TextFactory resolution and direct imports for both names.
